quote_agent/pyproject.toml
[project]
name = "quote_agent_simple"
version = "1.0.0"
description = "Simple A2A quote generator agent using OpenAI GPT models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "uvicorn>=0.34.3",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 